apps/api/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
python-multipart
google-generativeai
openai
requests
